I made this compile with ubuntu Linux gcc 4.3.2
Many char*s and char**s had to be changed to const char* and const char**.

For backwards compatibility there are two of these
	static int RunAllTests(int ac, const char** av);
	static int RunAllTests(int ac, char** av);

one to work with each possible main
	main(int argc, const char** argv).
	main(int argc, char** argv)

I did not want to break existing tests.


There are also two of thes in TestPlugin for the same reason.
	virtual bool parseAllArguments(int ac, const char** av, int index);
	virtual bool parseAllArguments(int ac, char** av, int index);

I did not modify the PluginTest file because parseArguments is not tested there.

I took out CHECK_EQUAL("THIS", "THIS") from places like
TEST(Utest, allMacros)
{
  CHECK(0 == 0);
  LONGS_EQUAL(1,1);
  BYTES_EQUAL(0xab,0xab);
  CHECK_EQUAL(100,100);
  STRCMP_EQUAL("THIS", "THIS");
  DOUBLES_EQUAL(1.0, 1.0, .01);
  POINTERS_EQUAL(this, this);
}

It was generating a warning.  For good reasons.  The use probably really wants STRCMP_EQUAL.

Changed Makefile CPPFLAGS += -Wall