updated readme file.
Added SRC_FILES symbol to makefile so that the ComponentMakefile can be very specific about which source files to include in a test run.

Fixed comment in release script.

Changed output of STRCMP_EQUAL and no case variety to put a marker "<!>" into a mismatched string pointing to the start of the mismatch.