Added a a _throw(except) and noexcept macro.
Doing throw(std::exception) is deprecated in C++11, but used for allocators
in C++03. Defining this _throw(except) macro makes it possible to add
throw(std::bad_alloc) in C++03, and nothing to C++11.

The noexcept macro is added in C++11 to replace the empty throw() to mean
we are not throwing any exceptions. using noexcept instead of throw() makes
us C++11 compatible.