Remove nothrow specifier hack for apple.
The __GLIBCXX_THROW defines itself as nothing for C++11, as the throw
specifiers are deprecated in C++11. This will not be changed to the same
type specifiers as is on C++03, and should be a problem for every C++11
compatible compiler.

A previous patch fixes this generally for C++11 compilers. Not only on
apple with glibc.