Adding call-checking when tracing is on.
Added an alternative tracing class, MockCheckedActualCallTrace, derived
from MockCheckedActualCall. (The class, MockActualCallTrace, is still
present, but its use with the tracing feature is disabled.) Access to a
few attributes of MockCheckedActualCall were given to the new class
through protected methods, to support the feature.

The MockSupport class was modified to use the new trace class in place
of the former one.

Tests in MockSupportTest.cpp were enhanced to cover the tracing feature
more completely.

A small correction to MockNamedValue (completing initialization) was
done; afaik this is not related to the tracing feature, but was found
during code analysis and appeared to be generally beneficial.