Juicy Bug Nr. 2 : Not resetting default comparator repository
Simply NOT setting it means it is left with a dangling pointer to
the previous tests stack. VC6 throws an exception for this. This
fixes it by resetting it to NULL.

TODO: Being static, this member should never be set to a local variable.
Even if this were not the case, each tests needs to return it to NULL
in teardown (or the so-called test "without" will not really be without).