Interface for checking an IEEE754 exception flag in the debugger
When a test fails due to an an IEEE754 exception flag being set,
it can be hard to find exactly where the error occurred.
By entering the new static method in the debugger as a watch,
e.g. IEEE754ExceptionsPlugin::checkIeee754ExeptionFlag(4), its
value will change to true at the exact point where the problem
occurred.