Basic implementation of "multi-matching" expectedCalls.
Now expectedCalls can match multiple actual calls, between a minimum and a maximum number of calls as defined by the user.

The method expectNCalls is not handled any more as just a collection of N expectedCall objects, but just a single expectedCall that requires exactly N calls to be considered fulfilled.

Other methods where also added to MockSupport to expect a range of calls (between a minimum and a maximum), at least a number of calls (i.e. minimum), at most a number of calls (i.e. maximum), and any number of calls.

Strict ordering checks are not working for the moment, except when all expected calls are for exactly 1 call (i.e. only using expectOneCall).

Actual calls order is not registered properly, so for the meantime failure messages just show the expectations that were fulfilled.

Composite calls are not needed any more, they have been disabled by pre-compiling them out (#if 0) along with their tests.