add support for long long integers as native types
When testing code that needs to work on both 64-bit platforms and 32-bit
platforms, the long integer may not guarantee 64bits of width on the
32-bit platform. This makes testing calls which take and return 64bit
values difficult, as the test framework can't handle these natively.

Add support for long long integers, so that test code can natively
handle these larger values.

Note, to avoid ABI compatibility issues, use cpputest_longlong and
cpputest_ulonglong. This ensures the virtual function tables are the
same whether long long support is enabled or disabled.

With this patch, if long long support is disabled in CppUTest, then the
mock withParameter and returnParameter functions will fail indicating
that the long long support is not enabled.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>