IAR: Add support to redirect file writes to stdout.
Ad the possibility to output strings that are normally written to a file to stdout. This makes it possible to generate JUnit XML output with the IAR simulator or even real HW. All FPuts are redirected to stdout and prepended with "FILEx" where x is an incrementing number. This makes it possible to separate the output later to different files.
Small caveat: All FPuts must be terminated with a newline character and should not contain newline characters except for the terminating one.