add .withUnmodifiedOutputParameter to expected call
When writing mock expectations, a mock implementation may have an output
parameter. In some tests it may be that the output parameter remains
unmodified.

To write an expectation which has an output parameter that is unmodified
you can use the following code:

    .withOutputParameterReturning("name", NULL, 0)

This is somewhat confusing, so add a shorthand

    .withUnmodifiedtOutputParameter("name")

This sets up the mock function to expect the "name" output parameter but
leave it unmodified.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>