Disable LIBNAME manipulation when used as subproject
`LIBNAME_POSTFIX_BITSIZE` is only useful when installing CppUTest,
not when consuming it with `add_subdirectory` or `FetchContent`.
(I'm not convinced it's justified even then; different architectures
should install to different _locations_ not names).

`LIBNAME_POSTFIX_DEBUG` is only useful when developing CppUTest,
not when consuming it.

Providing these options only serves to confuse the interface when users
shouldn't have to care.