Suppress Sanitization Warning on Linux When Memory Leak Detection is Disabled
There is a warning in the `CppUTestConfig.h` on Linux platforms when building
with ASAN. The warning says to disable memory leak detection in order to get
rid of the warning with `-DCPPUTEST_MEM_LEAK_DETECTION_DISABLED`. However,
providing said define does not get rid of the warning. This prevents
compilation on Linux with ASAN + `-Werror`.