add DEF_SOURCE_LINE to ctests
GoogleTest sets this property on ctest tests, and vscode-cmake-tools
will use it to dramatically improve the Test Explorer user experience.

CppUTest doesn't record file paths or line numbers for test groups, only
for tests, so when CPPUTEST_TESTS_DETAILED=0, use the lowest line number
of the tests in that group.